+++ src/utils/host.ts
@@ -34,3 +34,15 @@ export async function confirm(text: string, confirmLabel: string): Promise<boole
     const choice = await vscode.window.showWarningMessage(text, confirmLabel, 'Cancel');
     return choice === confirmLabel;
 }
+
+export async function showUnavailable(reason: "version-unknown" | "version-removed" | "extension-not-available") {
+    await vscode.window.showErrorMessage(unavailableMessage(reason));
+}
+
+function unavailableMessage(reason: "version-unknown" | "version-removed" | "extension-not-available"): string {
+    switch (reason) {
+        case "extension-not-available": return "Cannot run command: please check the 'Kubernetes' extension is installed";
+        case "version-removed": return "Cannot run command: please check for updates to the 'Open Policy Agent for Kubernetes' extension";
+        case "version-unknown": return "Cannot run command: please check for updates to the 'Kubernetes' extension";
+    }
+}
